New function
A new function to make a sub folder and move files there. Made by popular demand. Once the program has done with the files and moves them, a sub folder will be created with the date + time of move so if something goes wrong you know which files the promblem occured in